move SearchResponse classes to /classes dir and use protocol to define .run method searchers
