+++ weaviate_agents/query/classes/__init__.py
@@ -6,6 +6,7 @@ from .response import (
     AggregationResultWithCollection,
     AggregationResultWithCollectionNormalized,
     AskModeResponse,
+    AsyncSearchModeResponse,
     BooleanArrayPropertyFilter,
     BooleanMetrics,
     BooleanPropertyAggregation,
@@ -32,6 +33,7 @@ from .response import (
     QuerySort,
     QueryWithCollection,
     ResearchModeResponse,
+    SearchModeResponse,
     SearchModeResponseBase,
     Source,
     StreamedThoughts,
@@ -96,4 +98,6 @@ __all__ = [
     "UUIDArrayPropertyFilter",
     "SuggestedQuery",
     "SuggestQueryResponse",
+    "SearchModeResponse",
+    "AsyncSearchModeResponse",
 ]
