+++ weaviate_agents/query/classes/response.py
@@ -3,7 +3,16 @@ from __future__ import annotations
 import warnings
 from abc import ABC, abstractmethod
 from enum import Enum
-from typing import Any, Coroutine, Generic, Literal, Optional, TypeVar, Union
+from typing import (
+    Any,
+    Coroutine,
+    Generic,
+    Literal,
+    Optional,
+    Protocol,
+    TypeVar,
+    Union,
+)
 from uuid import UUID
 
 from pydantic import BaseModel, ConfigDict, Field, field_validator
@@ -525,3 +534,59 @@ class SearchModeResponseBase(BaseModel, ABC, Generic[SearcherT]):
         self: SearchModeResponseT, limit: int = 20, offset: int = 0
     ) -> Union[SearchModeResponseT, Coroutine[Any, Any, SearchModeResponseT]]:
         pass
+
+
+class _SyncSearcher(Protocol):
+    def run(self, limit: int = 20, offset: int = 0) -> SearchModeResponse: ...
+
+
+class _AsyncSearcher(Protocol):
+    async def run(
+        self, limit: int = 20, offset: int = 0
+    ) -> AsyncSearchModeResponse: ...
+
+
+class SearchModeResponse(SearchModeResponseBase[_SyncSearcher]):
+    """Response for the Query Agent search-only mode.
+
+    This contains the results of the search, the usage, and the underlying
+    searches performed. You can paginate through the results set by calling
+    the `next` method on this response with different `limit` / `offset` values.
+    This will result in the same underlying searches being performed each time,
+    resulting in a consistent results set across pages.
+    """
+
+    def next(self, limit: int = 20, offset: int = 0) -> SearchModeResponse:
+        """Paginate the search-only results with the given `limit` and `offset` values.
+
+        Args:
+            limit: The maximum number of results to return. If not specified, this defaults to 20.
+            offset: The offset to start from. If not specified, the retrieval begins from the first object in the results set.
+
+        Returns:
+            The next ``SearchModeResponse`` page.
+        """
+        return self._searcher.run(limit=limit, offset=offset)
+
+
+class AsyncSearchModeResponse(SearchModeResponseBase[_AsyncSearcher]):
+    """Response for the Query Agent search-only mode (async).
+
+    This contains the results of the search, the usage, and the underlying
+    searches performed. You can paginate through the results set by calling
+    the `next` method on this response with different `limit` / `offset` values.
+    This will result in the same underlying searches being performed each time,
+    resulting in a consistent results set across pages.
+    """
+
+    async def next(self, limit: int = 20, offset: int = 0) -> AsyncSearchModeResponse:
+        """Paginate the search-only results with the given `limit` and `offset` values.
+
+        Args:
+            limit: The maximum number of results to return. If not specified, this defaults to 20.
+            offset: The offset to start from. If not specified, the retrieval begins from the first object in the results set.
+
+        Returns:
+            The next ``AsyncSearchModeResponse`` page.
+        """
+        return await self._searcher.run(limit=limit, offset=offset)
